Solution to Check Greater Than

+++ Check Greater Than@@67882a731b4392d8fc1577d6/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 int main(int argc, char** argv) {
-   int a = strtol(argv[1], NULL, 10);
-   int b = strtol(argv[2], NULL, 10);
+   int a = (int)strtol(argv[1], NULL, 10);
+   int b = (int)strtol(argv[2], NULL, 10);
     a > b ? printf("True") : printf("False");
     return 0;
 }
